+++ sdk/communication/Azure.Communication.Identity/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 1.1.0-beta.1 (Unreleased)
-
+- Added ExchangeAccessToken operation
 
 ## 1.0.0 (2021-03-29)
 Updated `Azure.Communication.Identity` version.